Fix clang format script (#15)
There are two version of clang format: clang-format and git-clang-format. The former allows you to pass in files and it will just check the files. The latter, however, allows you to pass in a git commit SHA. It will then do a diff between the SHA you passed in and the HEAD (the most recent committed changes). It will then format those files, which means ultimately that it will only effect the parts of the files that were changed since the commit SHA.

What this means for us is that we need to pass in the SHA of the place we are merging into, which we assume to be origin/main***. So now in CI we do a fetch of origin/main, and the script will run against the last SHA of origin/main.

*** Note: This assumption might not always be valid, so ideally it would be best to check in a PR for the base of the PR (where we are merging back into) and get the SHA of it to pass in, but our solution works for now