Add `linux-x86_64` platform for bazel (#26)
This platform will allow for tools written in C that run locally on host machines that have a Linux operating system and an x86_64 CPU.

The Linux toolchain code in bazel/toolchain/linux.bzl is currently unused, and may be deleted later. The PR also introduces a constraint setting called environment, which will allow users to choose specific files to include or exclude based on where the code is running, whether it's on the HitL, MKV, MKVI, or in a test simulation.