Add manual signal controls to HitL CANController (#72)
Adds the ability to manually set signal values on the RKH CAN Bus through the set_state function in the CANController object. It also adds a test to test_can.py that uses two CANController objects to verify this functionality.

Also updates the testing .dbc files to scale throttle position from 0 to 100.