Add initial updated yaml->dbc generator
The new generator takes advantage of classes defined in the cantools
Python library for python abstractions of Messages and Signals. This
allows us to call the single `dump_file` function to dump our database
as a DBC file.

Additionally, it allows us to add in external DBC files like the one
from the motor controller, so we don't have to create our own YAML file
for the motor controller.