Add in BMS and AIR Control changes to MKV from comp (#229)
At competition, I added in a bunch of changes to the BMS and AIR Control code that never got merged. This includes modifying the state machine in the BMS to simplify it and cleaning up some of the air control code. It also adds support for 29-bit IDs in the CAN library, but no application code should need to use it.